% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/buildSpatialDataset.R
\name{buildSpatialDataset}
\alias{buildSpatialDataset}
\title{buildSpatialDataset}
\usage{
buildSpatialDataset(host, domain, cleanGeom = TRUE, unionStrategy = "bbox",
  verbose = TRUE, ids = NULL, exportPartialResults = FALSE,
  exportPath = getwd())
}
\arguments{
\item{host}{an object of class "character" giving the host}

\item{domain}{an object of class "character" giving the FIRMS domain}

\item{cleanGeom}{an object of class "logical" indicating if geometries have to
be validated with \pkg{cleangeo}. Default value is TRUE.}

\item{unionStrategy}{an object of class "character". Accepted values are "union"
(pure geoprocessing union - time consuming -), "bbox" (strategy to estimate the
largest spatial object to retain, by comparing envelopes, and results much less
time consuming). Default value is "bbox".}

\item{verbose}{an object of class "logical" either logs have to be printed out.
Default value is TRUE.}

\item{ids}{a vector of resource IDs for which the computation has to be run}

\item{exportPartialResults}{if partial shapefiles have to be exported Default
is FALSE.}

\item{exportPath}{the path were shapefiles should be exported}
}
\value{
an object of class "SpatialPolygonsDataFrame"
}
\description{
A function to build a spatial dataset for a given factsheet domain.
}
\note{
end-user function
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}

