% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/buildSpatialMetadata.R
\name{buildSpatialMetadata}
\alias{buildSpatialMetadata}
\title{buildSpatialMetadata}
\usage{
buildSpatialMetadata(sp)
}
\arguments{
\item{an}{object of class "SpatialPointsDataFrame}
}
\value{
an object of class "ISOMetadata"
}
\description{
A function to build a spatial metadata for the FIRMS data being processed
}
\note{
end-user function
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}

