% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSpatialObject.R
\name{buildSpatialObject}
\alias{buildSpatialObject}
\title{buildSpatialObject}
\usage{
buildSpatialObject(x, domain, wfs, verbose = TRUE)
}
\arguments{
\item{x}{FIRMS inventory item (as retrieved through web-services)}

\item{domain}{inventory domain (fishery/resource)}

\item{wfs}{an object of class "WFSClient" from \pkg{ows4R} package}

\item{verbose}{an object of class "logical" either logs have to be printed out.
Default value is TRUE.}
}
\value{
an object of class "SpatialPolygonsDataFrame"
}
\description{
A function to build a spatial object for a given factsheet
}
\note{
end-user function
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
