% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fetchFactsheetInfo.R
\name{fetchFactsheetAgencyInfo}
\alias{fetchFactsheetAgencyInfo}
\title{fetchFactsheetAgencyInfo}
\usage{
fetchFactsheetAgencyInfo(xml)
}
\arguments{
\item{xml}{an object of class "XmlInternalDocument"}
}
\value{
an object of class "character"
}
\description{
A function to fetch the owner acronym from the factsheet XML
}
\note{
function used internally to build FIRMS spatial objects
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}

