% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fetchFactsheetInfo.R
\name{fetchFactsheetIndicatorInfo}
\alias{fetchFactsheetIndicatorInfo}
\title{fetchFactsheetIndicatorInfo}
\usage{
fetchFactsheetIndicatorInfo(xml, indicator)
}
\arguments{
\item{xml}{an object of class "XmlInternalDocument"}

\item{indicator}{an object of class "character representing the xml indicator node}
}
\value{
an object of class "list"
}
\description{
A function to fetch statistical indicator information
}
\note{
function used internally to build FIRMS spatial objects
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}

