% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchFactsheetInfo.R
\name{fetchFactsheetInfo}
\alias{fetchFactsheetInfo}
\title{fetchFactsheetInfo}
\usage{
fetchFactsheetInfo(factsheet, lang, domain, host, species = FALSE,
  verbose = TRUE)
}
\arguments{
\item{factsheet}{an object of class "character" giving the factsheet FIGIS id}

\item{lang}{the language in which the factsheet is available}

\item{domain}{an object of class "character" giving the FIRMS domain}

\item{host}{an object of class "character" giving the host}

\item{species}{Include species information. Default is FALSE}

\item{verbose}{if logs should be printed out. Default is TRUE}
}
\value{
an object of class "list" giving the factsheet information
        (title and georef)
}
\description{
A function to fetch the factsheet information from a factsheet XML
}
\note{
function used internally to build FIRMS spatial objects
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
