% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchFactsheetInfo.R
\name{fetchFactsheetJurisdictionalInfo}
\alias{fetchFactsheetJurisdictionalInfo}
\title{fetchFactsheetJurisdictionalInfo}
\usage{
fetchFactsheetJurisdictionalInfo(xml)
}
\value{
an object of class "character"
}
\description{
A function to fetch the factsheet jurisdictional dist from a factsheet XML
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
