% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchFactsheetInfo.R
\name{fetchFactsheetReferences}
\alias{fetchFactsheetReferences}
\title{fetchFactsheetReferences}
\usage{
fetchFactsheetReferences(host, domain, verbose = TRUE)
}
\arguments{
\item{host}{an object of class "character" giving the host}

\item{domain}{an object of class "character" giving the FIRMS domain}

\item{verbose}{if logs should be printed out. Default is TRUE}
}
\value{
an object of class "data.frame"
}
\description{
A function to fetch the factsheet references (factsheet id, language)
}
\note{
function used internally to build FIRMS spatial objects
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
