% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchFactsheetInfo.R
\name{fetchFactsheetThematicApproachInfo}
\alias{fetchFactsheetThematicApproachInfo}
\title{fetchFactsheetThematicApproachInfo}
\usage{
fetchFactsheetThematicApproachInfo(xml)
}
\arguments{
\item{xml}{an object of class "XmlInternalDocument"}
}
\value{
an object of class "list"
}
\description{
A function to fetch the thematic approach information from the \code{ThematicApproach} 
node of a factsheet XML.
}
\note{
function used internally to build FIRMS spatial objects
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
