% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTabularTemplate.R
\name{getTabularTemplate}
\alias{getTabularTemplate}
\title{getTabularTemplate}
\usage{
getTabularTemplate(data, domain = c("fishery", "resource"))
}
\arguments{
\item{data}{a object of class \link{list} read from a JSON}

\item{domain}{the FIRMS domain either 'fishery' or 'resource'}
}
\value{
an object of class \link{data.frame} giving the expected tabular data structure
}
\description{
Produces a tabular data based on current FIRMS fishery/resource JSON web-service response
It aims to explain which field are used and expected to process FIRMS fishery/resource GIS placemarks used 
in the FIRMS stocks & fisheries map viewer.
This function is a temporary utility to illustrate the data structure that will be used in transition to dismiss 
the JSON approach followed now to fetch factsheet information.
}
\details{
getTabularTemplate
}
