% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prepareWFSRequest.R
\name{prepareWFSRequest}
\alias{prepareWFSRequest}
\title{prepareWFSRequest}
\usage{
prepareWFSRequest(layer)
}
\arguments{
\item{layer}{an object of class "data.frame" (single record) giving information
about a GIS web-layer to query}
}
\value{
an object of class "character" giving the WFS GetFeature request
}
\description{
A function to prepare the WFS request from a layer object
}
\note{
function used internally to build FIRMS spatial objects
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}

