% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readSpatialObject.R
\name{readSpatialObject}
\alias{readSpatialObject}
\title{readSpatialObject}
\usage{
readSpatialObject(item, cleanGeom = TRUE, verbose = TRUE)
}
\arguments{
\item{item}{an object of class "data.frame" (single record) giving information
about a GIS web-layer to query}

\item{cleanGeom}{an object of class "logical" indicating if geometries have to
be validated with \pkg{cleangeo}. Default value is TRUE.}

\item{verbose}{an object of class "logical" either logs have to be printed out.
Default value is TRUE.}
}
\value{
an object of class "data.frame" listing the GIS water area layer infos
}
\description{
A function to fetch the water area information from the \code{waterAreaRef}
node of a factsheet XML.
}
\note{
function used internally to build FIRMS spatial objects
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}

