% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFirmsGeo.R
\docType{package}
\name{RFirmsGeo}
\alias{RFirmsGeo-package}
\alias{RFirmsGeo}
\title{Tools for georeferencing FIRMS stocks and fisheries}
\description{
RFirmsGeo provides a set of facilities to georeference FIRMS stocks and fisheries
factsheet web-resources.
}
\details{
\tabular{ll}{
   Package: \tab RFirmsGeo\cr
   Type: \tab Package\cr
   Version: \tab 0.2-5\cr
   Date: \tab 2021-06-28\cr
   License: \tab GPL(>=2.0)\cr
   LazyLoad: \tab yes\cr
 }
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/openfigis/RFirmsGeo}
  \item Report bugs at \url{https://github.com/openfigis/RFirmsGeo/issues}
}

}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
