% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/buildGISLayerInfo.R
\name{buildGISLayerInfo}
\alias{buildGISLayerInfo}
\title{buildGISLayerInfo}
\usage{
buildGISLayerInfo(category, codesystem, code)
}
\arguments{
\item{category}{an object of class "character" describing the category
Possible values: "WaterArea", "SpeciesDistribution"}

\item{codesystem}{an object of class "character"}

\item{code}{an object of class "character"}
}
\value{
an object of class "list"
}
\description{
A utility function to build a GIS layer info, using \code{codesystem} and
\code{code} values inherited from a factsheet XML. The layer info consists in
a standard named list with all the information required to query a GIS layer, i.e.
the \code{url}, \code{typeName}, \code{propertyName}, \code{propertyValue}.

Additional fields, are added to report:
- the info \code{level} and the \code{rank}. Both fields are required to
optimize the query of FAO fishing areas, by using the finest fishing area
granularity.
- a \code{weight} specific to each layer. The weight is reversely proportional
to the geographic extent of the layer features. Maximum weight is attributed
to EEZs. In case of no overlap, or intersection failure, the feature that will
be kept to georeference a resource will be the one with the highest weight
(ie highest resolution).
}
\note{
function used internally to build FIRMS spatial objects
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}

