% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/buildSpatialDataset.R
\name{buildSpatialDataset}
\alias{buildSpatialDataset}
\title{buildSpatialDataset}
\usage{
buildSpatialDataset(host, domain, cleanGeom = TRUE, verbose = TRUE,
  ids = NULL, exportPartialResults = FALSE, exportPath = getwd())
}
\arguments{
\item{host}{an object of class "character" giving the host}

\item{domain}{an object of class "character" giving the FIRMS domain}

\item{cleanGeom}{an object of class "logical" indicating if geometries have to
be validated with \pkg{cleangeo}. Default value is TRUE.}

\item{verbose}{an object of class "logical" either logs have to be printed out.
Default value is TRUE.}

\item{ids}{a vector of indexes for which the computation has to be run}

\item{exportPartialResults}{if partial shapefiles have to be exported Default
is FALSE.}

\item{exportPath}{the path were shapefiles should be exported}
}
\value{
an object of class "SpatialPolygonsDataFrame"
}
\description{
A function to build a spatial dataset for a given factsheet domain.
}
\note{
end-user function
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}

