% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/buildSpatialObject.R
\name{buildSpatialObject}
\alias{buildSpatialObject}
\title{buildSpatialObject}
\usage{
buildSpatialObject(item, lang, host, domain, cleanGeom = TRUE,
  verbose = TRUE)
}
\arguments{
\item{item}{an object of class "character" giving the factsheet id}

\item{lang}{the language in which the factsheet is available}

\item{host}{an object of class "character" giving the host}

\item{domain}{an object of class "character" giving the FIRMS domain}

\item{cleanGeom}{an object of class "logical" indicating if geometries have to
be validated with \pkg{cleangeo}. Default value is TRUE.}

\item{verbose}{an object of class "logical" either logs have to be printed out.
Default value is TRUE.}
}
\value{
an object of class "SpatialPolygonsDataFrame"
}
\description{
A function to build a spatial object for a given factsheet
}
\note{
end-user function
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}

