% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fetchFactsheetInfo.R
\name{fetchFactsheetGeorefInfo}
\alias{fetchFactsheetGeorefInfo}
\title{fetchFactsheetGeorefInfo}
\usage{
fetchFactsheetGeorefInfo(xml, domain)
}
\arguments{
\item{xml}{an object of class "XmlInternalDocument"}

\item{domain}{an object of class "character representing the FIRMS resource}
}
\value{
an object of class "list"
}
\description{
A function to fetch the georeference information from the \code{Georeference}
node of a factsheet XML. Information is presented as named list, containing
the scale and title of the georeference.
}
\note{
function used internally to build FIRMS spatial objects
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}

