% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchFactsheetInfo.R
\name{fetchFactsheetInfo}
\alias{fetchFactsheetInfo}
\title{fetchFactsheetInfo}
\usage{
fetchFactsheetInfo(x, domain, verbose = TRUE)
}
\arguments{
\item{x}{inventory service item (as retrieved through web-service)}

\item{domain}{inventory domain (fishery/resource)}

\item{verbose}{if logs should be printed out. Default is TRUE}
}
\value{
an object of class "list" giving the factsheet information
        (title and georef)
}
\description{
A function to fetch the factsheet information from a factsheet XML
}
\note{
function used internally to build FIRMS spatial objects
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
