% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchFactsheetInfo.R
\name{fetchFactsheetSpeciesInfo}
\alias{fetchFactsheetSpeciesInfo}
\title{fetchFactsheetSpeciesInfo}
\usage{
fetchFactsheetSpeciesInfo(xml, domain)
}
\arguments{
\item{xml}{an object of class "XmlInternalDocument"}

\item{domain}{an object of class "character" the FIRMS domain}
}
\value{
an object of class "data.frame" listing the GIS species distributions 
        layer infos
}
\description{
A function to fetch the species information from the \code{SpeciesRef} node of
a factsheet XML.
}
\note{
function used internally to build FIRMS spatial objects
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
