% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSpatialObject.R
\name{readSpatialObject}
\alias{readSpatialObject}
\title{readSpatialObject}
\usage{
readSpatialObject(wfs, layer, cleanGeom = TRUE,
  cleanStrategy = "BUFFER", verbose = TRUE)
}
\arguments{
\item{wfs}{instance of WFS client, object of class \code{WFSClient} from \pkg{ows4R}}

\item{layer}{an object of class "data.frame" (single record) giving information
about a GIS web-layer to query}

\item{cleanGeom}{an object of class "logical" indicating if geometries have to
be validated with \pkg{cleangeo}. Default value is TRUE.}

\item{cleanStrategy}{an object of class "character". Default value is "BUFFER"}

\item{verbose}{an object of class "logical" either logs have to be printed out.
Default value is TRUE.}
}
\value{
an object of class "spatial" giving the spatial object
}
\description{
A function to read a spatial object from an layer item
}
\note{
function used internally to build FIRMS spatial objects
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
