% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readSpatialObject.R
\name{readSpatialObjects}
\alias{readSpatialObjects}
\title{readSpatialObjects}
\usage{
readSpatialObjects(layers, cleanGeom, verbose)
}
\arguments{
\item{layers}{an object of class "data.frame" (single record) giving information
about a GIS web-layer to query}

\item{cleanGeom}{an object of class "logical" indicating if geometries have to
be validated with \pkg{cleangeo}. Default value is TRUE.}

\item{verbose}{an object of class "logical" either logs have to be printed out.
Default value is TRUE.}
}
\value{
an object of class "list" listing the spatial objects
}
\description{
A function to list spatial objects from a list of layer items
}
\note{
function used internally to build FIRMS spatial objects
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}

