% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RFirmsGeo.R
\docType{package}
\name{RFirmsGeo}
\alias{RFirmsGeo}
\alias{RFirmsGeo-package}
\title{Tools for georeferencing FIRMS stocks and fisheries}
\description{
RFirmsGeo provides a set of facilities to georeference FIRMS stocks and fisheries
factsheet web-resources.
}
\details{
\tabular{ll}{
   Package: \tab RFirmsGeo\cr
   Type: \tab Package\cr
   Version: \tab 0.1-0\cr
   Date: \tab 2015-10-20\cr
   License: \tab GPL(>=2.0)\cr
   LazyLoad: \tab yes\cr
 }
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}

